[ClientController] implement GetAll and Create on ClientController

+++ RentallCarsAPI/Models/Client.cs
@@ -15,7 +15,7 @@ namespace RentallCarsAPI.Models
         public string Address { get; set; }
         public string City { get; set; }
         public string Province { get; set; }
-        public int PostalCode { get; set; }
+        public short PostalCode { get; set; }
         public DateTime LastModification { get; set; }
     }
 }
